*New features
At the end of the quiz, a message will tell the user that the quiz has ended

*Optimization

Changed algorithm for testing the correct answers when user has pressed the check answer button
Now it will first search if the users ansers matches that of the questions correct answers
If that's the case return true.

Optimized code by reducing the time the app needs to find DOM elements, by introducing
DOM variables to respective state. As an example the member variable listOfAnswers will now point to
the unordered list of input answers

Removed unused HTML ids, CSS classes and JS functions

*Bugfixing

Fixed an issue were the retry button and settings were adjusting heights
when switching to intro state and questionstate. However, theres still
the same error when switching to result state